requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.1
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
steam==1.4.4
mal-api==0.5.1
tweepy==4.14.0
spotipy==2.23.0
PyGitHub==2.3.0
schedule==1.2.1
